Add support for assistance type verification
